# docs/api-spec.yaml
openapi: 3.1.0
info:
  title: MockAgents API
  version: 0.4.0
  description: >
    MockAgents is a platform for mocking AI agent integrations. It exposes a
    Management API (agents, interaction logs, costs, audit, pipelines, config
    validation, and a multi-tenant control plane of tenants + RBAC API keys),
    and Protocol Endpoints that mimic real LLM provider APIs (OpenAI Chat
    Completions and Anthropic Messages).


    The Model Context Protocol mock (`/mcp`, `/mcp/events`, `/mcp/response`) is
    JSON-RPC 2.0 rather than REST and is documented in the project README,
    not in this OpenAPI document.


    **Authentication.** When the server runs with `MOCKAGENTS_MULTI_TENANT=1`,
    every `/api/v1/*` management route requires an API key
    (`Authorization: Bearer <key>` or `X-Api-Key: <key>`) whose role meets the
    route's floor (`viewer < editor < admin < platform`). In single-tenant mode
    the management API is open. The protocol endpoints (`/v1/*`) are always
    unauthenticated — clients send their own provider key, which MockAgents
    ignores.
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /v1/chat/completions:
    post:
      operationId: openaiChatCompletions
      summary: OpenAI-compatible Chat Completions
      description: >
        Accepts requests in the OpenAI Chat Completions format, matches the
        model name to a loaded mock agent, and returns a mock response that
        follows the OpenAI response schema. Supports both synchronous JSON
        responses and streaming via Server-Sent Events.
      tags: [OpenAI]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OpenAIChatCompletionRequest"
      responses:
        "200":
          description: >
            Chat completion response. When `stream` is false (or omitted) the
            response is a single JSON object. When `stream` is true the
            response is a stream of Server-Sent Events with `Content-Type:
            text/event-stream`; each event carries a
            `OpenAIChatCompletionChunk` JSON payload, and the stream ends with
            a `data: [DONE]` sentinel.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OpenAIChatCompletionResponse"
            text/event-stream:
              schema:
                description: >
                  SSE stream of `OpenAIChatCompletionChunk` objects. Each
                  event is prefixed with `data: ` followed by a JSON line.
                  The stream terminates with `data: [DONE]`.
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"
  /v1/messages:
    post:
      operationId: anthropicMessages
      summary: Anthropic-compatible Messages
      description: >
        Accepts requests in the Anthropic Messages format, matches the model
        name to a loaded mock agent, and returns a mock response that follows
        the Anthropic response schema. Supports both synchronous JSON
        responses and streaming via Server-Sent Events with Anthropic event
        types (message_start, content_block_start, content_block_delta,
        content_block_stop, message_delta, message_stop).
      tags: [Anthropic]
      parameters:
        - name: x-api-key
          in: header
          description: Anthropic API key (accepted but not validated by the mock).
          schema:
            type: string
        - name: anthropic-version
          in: header
          description: Anthropic API version header.
          schema:
            type: string
            example: "2023-06-01"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnthropicMessagesRequest"
      responses:
        "200":
          description: >
            Messages response. When `stream` is false (or omitted) the response
            is a single JSON object. When `stream` is true the response is a
            stream of Server-Sent Events with `Content-Type: text/event-stream`;
            each event has an `event:` type field (message_start,
            content_block_start, content_block_delta, content_block_stop,
            message_delta, message_stop) and a `data:` JSON payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnthropicMessagesResponse"
            text/event-stream:
              schema:
                description: >
                  SSE stream using Anthropic event types. Each frame consists
                  of an `event:` line and a `data:` JSON line.
                type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalServerError"

# =============================================================================
# Components
# =============================================================================
components:
...
  schemas:
...
    Error:
      type: object
      description: >
        The management API's error envelope. Every 4xx and 5xx from a
        /api/v1/... route has this shape.

        NOT what the provider endpoints return when a quota rejects a request:
        those mimic the upstream `{"error": {"type", "message"}}` object so an
        SDK surfaces them correctly. See ProviderQuotaError.
      required: [error]
      properties:
        error:
          type: string
          example: agent "weather-agent" not found
        available_agents:
          type: array
          items: { type: string }
          description: >
            Present on a 404 for an unknown agent — what the caller could have
            asked for instead. Tenant-scoped: it never names another tenant's
            agents.
        details:
          type: string
          description: >
            Present when a write or reload failed schema validation; carries the
            validator's output.
...
    OpenAIChatCompletionRequest:
      type: object
      description: >
        Sampling parameters the real API honours — `top_p`, `n`, `stop`,
        `user`, `frequency_penalty`, `presence_penalty` — and a message's
        `name` are accepted and ignored here: responses come from the agent's
        scenarios, not from sampling. They are documented because a client may
        send them and will not get an error, not because they do anything.
      required: [model, messages]
      properties:
        parallel_tool_calls:
          type: boolean
          description: >
            Honoured under the strict-tools knob, which caps how many
            simultaneous tool calls a scenario may return.
        response_format:
          type: object
          additionalProperties: true
          description: Structured-output request; json_schema mode is honoured.
        stream_options:
          type: object
          additionalProperties: true
          description: >
            `include_usage` adds a final usage-only chunk to a stream, as the
            real API does.
        model:
          type: string
          description: >
            Model identifier. MockAgents uses this to resolve the target mock
            agent.
          example: weather-agent
        messages:
          type: array
          items:
            $ref: "#/components/schemas/OpenAIChatMessage"
          minItems: 1
        tools:
          type: array
          items:
            $ref: "#/components/schemas/OpenAITool"
        tool_choice:
          description: >
            Controls which (if any) tool is called. Can be "none", "auto",
            "required", or an object specifying a particular function.
            "none" is always honored; "required" and named forcing (with
            forced-call synthesis and finish_reason "stop") are enforced
            when the agent's strict_tools knob is on — otherwise they are
            accepted and ignored.
          oneOf:
            - type: string
              enum: [none, auto, required]
            - type: object
              required: [type, function]
              properties:
                type:
                  type: string
                  enum: [function]
                function:
                  type: object
                  required: [name]
                  properties:
                    name:
                      type: string
        stream:
          type: boolean
          default: false
          description: If true, partial message deltas are sent as SSE events.
        temperature:
          type: number
          minimum: 0
          maximum: 2
          default: 1
        max_tokens:
          type: integer
          minimum: 1
        top_p:
          type: number
          minimum: 0
          maximum: 1
        frequency_penalty:
          type: number
          minimum: -2
          maximum: 2
        presence_penalty:
          type: number
          minimum: -2
          maximum: 2
        stop:
          description: Up to 4 sequences where the API will stop generating.
          oneOf:
            - type: string
            - type: array
              items:
                type: string
              maxItems: 4
        n:
          type: integer
          minimum: 1
          default: 1
        user:
          type: string
    OpenAIChatMessage:
      type: object
      required: [role]
      properties:
        role:
          type: string
          enum: [system, user, assistant, tool]
        content:
          description: Message content; string or array of content parts.
          oneOf:
            - type: string
            - type: array
              items:
                type: object
                properties:
                  type:
                    type: string
                    enum: [text, image_url]
                  text:
                    type: string
                  image_url:
                    type: object
                    properties:
                      url:
                        type: string
                        format: uri
        name:
          type: string
        tool_calls:
          type: array
          items:
            $ref: "#/components/schemas/OpenAIToolCall"
        tool_call_id:
          type: string
          description: Required when role is "tool".
    OpenAITool:
      type: object
      required: [type, function]
      properties:
        type:
          type: string
          enum: [function]
        function:
          type: object
          required: [name]
          properties:
            name:
              type: string
            description:
              type: string
            parameters:
              type: object
              additionalProperties: true
              description: JSON Schema describing the function parameters.
    OpenAIToolCall:
      type: object
      required: [id, type, function]
      properties:
        id:
          type: string
        type:
          type: string
          enum: [function]
        function:
          type: object
          required: [name, arguments]
          properties:
            name:
              type: string
            arguments:
              type: string
              description: JSON-encoded arguments string.
    OpenAIChatCompletionResponse:
      type: object
      required: [id, object, created, model, choices, usage]
      properties:
        id:
          type: string
          example: chatcmpl-abc123
        object:
          type: string
          enum: [chat.completion]
        created:
          type: integer
          description: Unix timestamp.
        model:
          type: string
        choices:
          type: array
          items:
            type: object
            required: [index, message, finish_reason]
            properties:
              index:
                type: integer
              message:
                $ref: "#/components/schemas/OpenAIChatMessage"
              finish_reason:
                type: string
                enum: [stop, length, tool_calls, content_filter]
        usage:
          $ref: "#/components/schemas/OpenAIUsage"
    OpenAIUsage:
      type: object
      required: [prompt_tokens, completion_tokens, total_tokens]
      properties:
        prompt_tokens:
          type: integer
        completion_tokens:
          type: integer
        total_tokens:
          type: integer
...
    AnthropicMessagesRequest:
      type: object
      description: >
        Sampling parameters the real API honours — `temperature`, `top_p`,
        `top_k`, `stop_sequences` — and `metadata` are accepted and ignored
        here: responses come from the agent's scenarios, not from sampling.
        They are documented because a client may send them and will not get an
        error, not because they do anything.
      required: [model, messages, max_tokens]
      properties:
        model:
          type: string
          description: >
            Model identifier. MockAgents uses this to resolve the target mock
            agent.
          example: weather-agent
        messages:
          type: array
          items:
            $ref: "#/components/schemas/AnthropicMessage"
          minItems: 1
        system:
          description: System prompt; string or array of content blocks.
          oneOf:
            - type: string
            - type: array
              items:
                type: object
                properties:
                  type:
                    type: string
                    enum: [text]
                  text:
                    type: string
        tools:
          type: array
          items:
            $ref: "#/components/schemas/AnthropicTool"
        thinking:
          type: object
          description: >
            Extended-thinking gate. `type: enabled` makes the response carry a
            synthesized thinking block; `budget_tokens` is advisory.
          properties:
            type: { type: string, enum: [enabled] }
            budget_tokens: { type: integer }
        tool_choice:
          type: object
          description: >
            "none" is always honored; "any"/"tool" forcing and
            disable_parallel_tool_use are enforced when the agent's
            strict_tools knob is on.
          properties:
            type:
              type: string
              enum: [auto, any, tool, none]
            name:
              type: string
              description: Required when type is "tool".
            disable_parallel_tool_use:
              type: boolean
              description: Cap the response to a single tool_use block.
        max_tokens:
          type: integer
          minimum: 1
          example: 1024
        stream:
          type: boolean
          default: false
        temperature:
          type: number
          minimum: 0
          maximum: 1
        top_p:
          type: number
          minimum: 0
          maximum: 1
        top_k:
          type: integer
          minimum: 0
        stop_sequences:
          type: array
          items:
            type: string
        metadata:
          type: object
          properties:
            user_id:
              type: string
    AnthropicMessage:
      type: object
      required: [role, content]
      properties:
        role:
          type: string
          enum: [user, assistant]
        content:
          description: Message content; string or array of content blocks.
          oneOf:
            - type: string
            - type: array
              items:
                $ref: "#/components/schemas/AnthropicRequestContentBlock"
    AnthropicContentBlock:
      type: object
      description: >
        A content block the SERVER emits — in a response, or in a
        content_block_start stream event. Request-side blocks a client may send
        are a wider set (tool_result, image) and are accepted as free-form; see
        AnthropicRequestContentBlock.
      required: [type]
      properties:
        type:
          type: string
          enum: [text, thinking, tool_use]
        text:
          type: string
          description: Present on a text block.
        thinking:
          type: string
          description: >
            Present on a thinking block, emitted when the request enabled
            extended thinking.
        signature:
          type: string
          description: Opaque signature accompanying a thinking block.
        id:
          type: string
          description: Present on a tool_use block.
        name:
          type: string
          description: Present on a tool_use block.
        input:
          type: object
          additionalProperties: true
          description: >
            A tool_use block's arguments. Always present on such a block, as
            `{}` for a no-argument call — strict SDK validation rejects a
            missing key.
    AnthropicRequestContentBlock:
      type: object
      description: >
        A content block a CLIENT may send. Accepted as free-form: the mock reads
        the parts it matches on (text, and the presence of images) and passes
        the rest through without modelling it, so this documents the shape
        rather than constraining it.
      additionalProperties: true
      properties:
        type:
          type: string
          enum: [text, image, tool_use, tool_result]
        text: { type: string }
        source:
          type: object
          additionalProperties: true
          description: An image block's data. Counted for has_image matching.
        tool_use_id: { type: string, description: Present on a tool_result block. }
        content:
          description: A tool_result block's payload; string or array of blocks.
        is_error: { type: boolean, description: Marks a tool_result as an error. }
    AnthropicTool:
      type: object
      required: [name, input_schema]
      properties:
        name:
          type: string
        description:
          type: string
        input_schema:
          type: object
          additionalProperties: true
          description: JSON Schema describing the tool input.
        cache_control:
          type: object
          additionalProperties: true
          description: >
            Prompt-cache breakpoint. Honoured: a first request carrying one
            bills cache_creation_input_tokens, an identical repeat bills
            cache_read_input_tokens.
    AnthropicMessagesResponse:
      type: object
      required: [id, type, role, content, model, stop_reason, usage]
      properties:
        id:
          type: string
          example: msg_abc123
        type:
          type: string
          enum: [message]
        role:
          type: string
          enum: [assistant]
        content:
          type: array
          items:
            $ref: "#/components/schemas/AnthropicContentBlock"
        model:
          type: string
        stop_reason:
          type:
            - string
            - "null"
          enum: [end_turn, max_tokens, stop_sequence, tool_use, null]
        stop_sequence:
          type:
            - string
            - "null"
        usage:
          $ref: "#/components/schemas/AnthropicUsage"
    AnthropicUsage:
      type: object
      required: [input_tokens, output_tokens]
      properties:
        input_tokens:
          type: integer
        output_tokens:
          type: integer
        cache_creation_input_tokens:
          type: integer
          nullable: true
          description: >
            Present only when the request carried cache_control markers, and
            then as a MATCHED PAIR with cache_read_input_tokens — one side may
            be 0. Matches the real API, where both appear together.
        cache_read_input_tokens:
          type: integer
          nullable: true
          description: See cache_creation_input_tokens; the two appear together.
...
  responses:
    BadRequest:
      description: The request was malformed or contained invalid parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: bad_request
              message: "Invalid value for parameter 'limit': must be a positive integer."
              details: {}
    NotFound:
      description: The requested resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: not_found
              message: The requested resource was not found.
              details: {}
    InternalServerError:
      description: An unexpected server error occurred.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error:
              code: internal_error
              message: An unexpected error occurred.
              details: {}
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: >
        `Authorization: Bearer <api-key>`. Required on `/api/v1/*` routes when
        the server runs with `MOCKAGENTS_MULTI_TENANT=1`.
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: "Alternative to `Authorization: Bearer` for the same API key."

# Applies to the management API in multi-tenant mode. Protocol endpoints
# (/v1/*) are always unauthenticated and override this with `security: []`.
security:
  - BearerAuth: []
  - ApiKeyAuth: []